Shuffle original array, reset currentQuestionIndex and export it
The reason as to why the original array needs to be shuffled is because
that the same question will reappear in the same order when replay is pressed.
That is because on page load, the displayQuestion function is executed and
and another array is created which will always remain the same.
Adding shuffleArray(quizQuestions) will always shuffle the original array
making it never appear in the same order. Also reset the currentQuestionIndex
so the replay function works and it starts over from 0 to 10.